Added types file
