+++ src/index.html
@@ -28,20 +28,20 @@
 
     <script>
       var config = {
-  apiEndpoint: "https://api.g8s.gauss.eu-central-1.aws.gigantic.io",
-  // apiEndpoint: "https://api.g8s.seal.us-east-1.aws.gigantic.io",
-  // apiEndpoint: "https://gs-api.g8s.eu-central-1.aws.cps.vodafone.com",
-  // apiEndpoint: "https://api.g8s.gorilla.eu-central-1.aws.gigantic.io",
-  // apiEndpoint: "https://api.g8s.godsmack.westeurope.azure.gigantic.io",
-  // apiEndpoitn: "https://api.g8s.geckon.gridscale.kvm.gigantic.io",
-  passageEndpoint: "http://localhost:5001",
-  environment: "development",
-  ingressBaseDomain: "k8s.sample.io",
-  awsCapabilitiesJSON:
-    '{"c5.2xlarge":{"cpu_cores":8,"description":"C5 Compute Optimized Double Extra Large","memory_size_gb":16,"storage_size_gb":0},"c5.4xlarge":{"cpu_cores":16,"description":"C5 Compute Optimized Quadruple Extra Large","memory_size_gb":32,"storage_size_gb":0},"c5.9xlarge":{"cpu_cores":36,"description":"C5 Compute Optimized Nonuple Extra Large","memory_size_gb":72,"storage_size_gb":0},"c5.large":{"cpu_cores":2,"description":"C5 Compute Optimized Large","memory_size_gb":4,"storage_size_gb":0},"c5.xlarge":{"cpu_cores":4,"description":"C5 Compute Optimized Extra Large","memory_size_gb":8,"storage_size_gb":0},"c5n.4xlarge":{"cpu_cores":16,"description":"C5N Compute Optimized Quadruple Extra Large","memory_size_gb":42,"storage_size_gb":0},"i3.xlarge":{"cpu_cores":4,"description":"I3 Storage Optimized Extra Large","memory_size_gb":30.5,"storage_size_gb":950},"m3.2xlarge":{"cpu_cores":8,"description":"M3 General Purpose Double Extra Large","memory_size_gb":30,"storage_size_gb":80},"m3.large":{"cpu_cores":2,"description":"M3 General Purpose Large","memory_size_gb":7.5,"storage_size_gb":32},"m3.xlarge":{"cpu_cores":4,"description":"M3 General Purpose Extra Large","memory_size_gb":15,"storage_size_gb":40},"m4.2xlarge":{"cpu_cores":8,"description":"M4 General Purpose Double Extra Large","memory_size_gb":32,"storage_size_gb":0},"m4.4xlarge":{"cpu_cores":16,"description":"M4 General Purpose Four Extra Large","memory_size_gb":64,"storage_size_gb":0},"m4.large":{"cpu_cores":2,"description":"M4 General Purpose Large","memory_size_gb":8,"storage_size_gb":0},"m4.xlarge":{"cpu_cores":4,"description":"M4 General Purpose Extra Large","memory_size_gb":16,"storage_size_gb":0},"m5.2xlarge":{"cpu_cores":8,"description":"M5 General Purpose Double Extra Large","memory_size_gb":32,"storage_size_gb":0},"m5.4xlarge":{"cpu_cores":16,"description":"M5 General Purpose Quadruple Extra Large","memory_size_gb":64,"storage_size_gb":0},"m5.large":{"cpu_cores":2,"description":"M5 General Purpose Large","memory_size_gb":8,"storage_size_gb":0},"m5.xlarge":{"cpu_cores":4,"description":"M5 General Purpose Extra Large","memory_size_gb":16,"storage_size_gb":0},"p2.xlarge":{"cpu_cores":4,"description":"P2 Extra Large providing GPUs","memory_size_gb":61,"storage_size_gb":0},"p3.2xlarge":{"cpu_cores":8,"description":"P3 Double Extra Large providing GPUs","memory_size_gb":61,"storage_size_gb":0},"p3.8xlarge":{"cpu_cores":32,"description":"P3 Eight Extra Large providing GPUs","memory_size_gb":244,"storage_size_gb":0},"r3.2xlarge":{"cpu_cores":8,"description":"R3 High-Memory Double Extra Large","memory_size_gb":61,"storage_size_gb":160},"r3.4xlarge":{"cpu_cores":16,"description":"R3 High-Memory Quadruple Extra Large","memory_size_gb":122,"storage_size_gb":320},"r3.8xlarge":{"cpu_cores":32,"description":"R3 High-Memory Eight Extra Large","memory_size_gb":244,"storage_size_gb":320},"r3.large":{"cpu_cores":2,"description":"R3 High-Memory Large","memory_size_gb":15.25,"storage_size_gb":32},"r3.xlarge":{"cpu_cores":4,"description":"R3 High-Memory Extra Large","memory_size_gb":30.5,"storage_size_gb":80},"r5.2xlarge":{"cpu_cores":8,"description":"R5 High-Memory Double Extra Large","memory_size_gb":64,"storage_size_gb":0},"t2.2xlarge":{"cpu_cores":8,"description":"T2 General Purpose Double Extra Large","memory_size_gb":32,"storage_size_gb":0},"t2.large":{"cpu_cores":2,"description":"T2 General Purpose Large","memory_size_gb":8,"storage_size_gb":0},"t2.xlarge":{"cpu_cores":4,"description":"T2 General Purpose Extra Large","memory_size_gb":16,"storage_size_gb":0}}',
-  azureCapabilitiesJSON:
-    '{"Standard_A2_v2":{"additionalProperties":{},"description":"This is some description","maxDataDiskCount":4,"memoryInMb":4294.967296,"name":"Standard_A2_v2","numberOfCores":2,"osDiskSizeInMb":1047552,"resourceDiskSizeInMb":21474.83648},"Standard_A4_v2":{"additionalProperties":{},"description":"Here is a longer description that might be too long for the field","maxDataDiskCount":8,"memoryInMb":8589.934592,"name":"Standard_A4_v2","numberOfCores":4,"osDiskSizeInMb":1047552,"resourceDiskSizeInMb":42949.67296},"Standard_A8_v2":{"additionalProperties":{},"description":"Another VM size description text","maxDataDiskCount":16,"memoryInMb":17179.869184,"name":"Standard_A8_v2","numberOfCores":8,"osDiskSizeInMb":1047552,"resourceDiskSizeInMb":85899.34592}, "Standard_D2s_v3":{"additionalProperties":{},"description":"Dsv3-series, general purpose, 160-190 ACU, premium storage supported","maxDataDiskCount":4,"memoryInMb":8589.935,"name":"Standard_D2s_v3","numberOfCores":2,"osDiskSizeInMb":1047552,"resourceDiskSizeInMb":17179.87}}'
-};
+        apiEndpoint: 'http://localhost:8000',
+        passageEndpoint: 'http://localhost:5001',
+        // environment is set to:
+        // 'dev' while developing
+        // 'kubernetes' when deployed in production
+        // 'docker-container' when the final build container is run manually
+        // the behaviour is configured by the ENVIRONMENT variable,
+        // see scripts/start.sh
+        environment: 'development',
+        ingressBaseDomain: 'k8s.sample.io',
+        awsCapabilitiesJSON: '{"m3.large":{"description":"M3 General Purpose Large","memory_size_gb":7.5,"cpu_cores":2,"storage_size_gb":32},"m3.xlarge":{"description":"M3 General Purpose Extra Large","memory_size_gb":15,"cpu_cores":4,"storage_size_gb":80},"m3.2xlarge":{"description":"M3 General Purpose Double Extra Large","memory_size_gb":30,"cpu_cores":8,"storage_size_gb":160}, "m4.xlarge":{"cpu_cores":4,"description":"M4 General Purpose Extra Large","memory_size_gb":16,"storage_size_gb":0}}',
+        azureCapabilitiesJSON: '{"Standard_A2_v2":{"additionalProperties":{},"description":"This is some description","maxDataDiskCount":4,"memoryInMb":4294.967296,"name":"Standard_A2_v2","numberOfCores":2,"osDiskSizeInMb":1047552,"resourceDiskSizeInMb":21474.83648},"Standard_A4_v2":{"additionalProperties":{},"description":"Here is a longer description that might be too long for the field","maxDataDiskCount":8,"memoryInMb":8589.934592,"name":"Standard_A4_v2","numberOfCores":4,"osDiskSizeInMb":1047552,"resourceDiskSizeInMb":42949.67296},"Standard_A8_v2":{"additionalProperties":{},"description":"Another VM size description text","maxDataDiskCount":16,"memoryInMb":17179.869184,"name":"Standard_A8_v2","numberOfCores":8,"osDiskSizeInMb":1047552,"resourceDiskSizeInMb":85899.34592}, "Standard_D2s_v3":{"additionalProperties":{},"description":"Dsv3-series, general purpose, 160-190 ACU, premium storage supported","maxDataDiskCount":4,"memoryInMb":8589.935,"name":"Standard_D2s_v3","numberOfCores":2,"osDiskSizeInMb":1047552,"resourceDiskSizeInMb":17179.87}}',
+        happaVersion: 'development',
+      };
       window.config = config;
     </script>
